server/requirements.txt
ultralytics>=8.3.0
fastapi>=0.115.0
uvicorn>=0.34.0
python-multipart>=0.0.18
pyyaml>=6.0
onnx>=1.17.0
onnxruntime-gpu>=1.20.0
Pillow>=10.0.0
